LayoutTests/fast/history/history-back-initial-vs-final-url-expected.txt

ALERT: Initial URL loaded.
ALERT: Final URL loaded.
ALERT: Going back.
ALERT: Final URL loaded.
Checks that when going back to a page that frames that were navigated, that the final URL of the frames is loaded directly, instead the initial URL. Four alerts should appear, in this order:

Initial URL loaded.
Final URL loaded.
Going back.
Final URL loaded.


--------
Frame: '<!--frame1-->'
--------
final page contents

============== Back Forward List ==============
curr->  (file test):fast/history/history-back-initial-vs-final-url.html  **nav target**
            (file test):fast/history/resources/frame-final-url.html (in frame "<!--frame1-->")
        data:text/html,<script>alert("Going back.");history.back();</script>  **nav target**
===============================================
